[Part 02] Exception handling

+++ src/main/java/com/texus/springboot/web/springbootfirstwebapplication/controller/TodoController.java
@@ -54,6 +54,8 @@ public class TodoController {
 
     @RequestMapping(value="/delete-todo", method = RequestMethod.GET)
     public String deleteTodo(@RequestParam int id) {
+        if(id == 1)
+            throw new RuntimeException("Something went wrong!");
         service.deleteTodo(id);
         return "redirect:/list-todos";
     }
